Fixes and questions
Alright, so the "everyone spawns when I spawn" issue is solved. Now the question is, is everyone sending and receiving messages to the server? I'm thinking it might not be working because I never told the characters to spawn. It simply sends a reference to the parts and doesn't tell the character assembler to work. I'd fix it right now but FUCKING KPIS.