Player Sync
Working on the Player Class. WE NEED A WAY TO TRACK THE CORRECT ID FOR EACH PLAYER. MAYBE MAKE THE MASTER CREATE THE IDS FOR EACH PLAYER, THEN SEND IT TO EACH ONE OF THEM. SEND THE NEW NAME TO THE OTHER CLIENTS. After that, each client should compare the new id with theirs and detect if they're the correct client to detect input, THEN create the Controller object with id 0.