More Fixes
Fixed PlayerScoreController and PlayerUIModule

To do:
- Fix weapon control initialization (replace START with something else)
- Fix defensive skill control initialization (replace START with something else)
- Fix complementary control initialization

We might need newer callbacks for the three classes, making the player initialize them instead of letting the game manager take full control. But how the fuck do I do it? IDK. Maybe assign layers to them and then make the player search for the components with the right layer when initializing.